requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
PyJWT
passlib[argon2]
argon2-cffi
jinja2
python-multipart